backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
httpx==0.25.1
firebase-admin==6.2.0
google-cloud-aiplatform==1.38.1
google-cloud-documentai==2.20.1
ortools==9.8.3296
networkx==3.2.1
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
pillow==10.1.0
reportlab==4.0.7
python-dotenv==1.0.0
